+++ classes/local/checks/environ.php
@@ -54,7 +54,7 @@ class environ extends check_base {
     public function success(): bool {
         $maxexectime = $this->model->get_details()['max_execution_time'];
         return $this->model->status === constants::STATUS_FINISHED
-            && (!$maxexectime || $maxexectime >= HOURSECS);
+            && (!$maxexectime || $maxexectime >= constants::MIN_REQUIRED_EXEC_TIME);
     }
 
     /**
@@ -62,7 +62,7 @@ class environ extends check_base {
      *
      * @return bool
      */
-    protected function warning(): bool {
+    public function warning(): bool {
         return $this->success() && !empty($this->model->get_details()['max_execution_time']);
     }